fix:variable names fix,simplified prop drilling
