+++ src/App.tsx
@@ -1,22 +1,26 @@
 import React, { useState } from 'react';
 import './App.css';
 import Title from './components/Title';
-import { AddItemForm, TaskListMap } from './components/AddItemForm';
+import { AddItemForm, TaskType } from './components/AddItemForm';
 import TaskDashboard from './components/TaskDashboard';
 
 function App() {
-  const [userInput, setUserInput] = useState<string>('');
-  const [taskList, setTaskList] = useState<TaskListMap[]>([
+  const [taskList, setTaskList] = useState<TaskType[]>([
     { task: 'Model Task Example', completed: false, id: '123456' },
   ]);
+  function addNewTask(taskContent: string) {
+    const myNewTask = {
+      task: taskContent,
+      completed: false,
+      id: String(new Date().valueOf()),
+    };
+    setTaskList([...taskList, myNewTask]);
+  }
   return (
     <div className="todoapp stack-large">
       <Title />
       <AddItemForm
-        userInput={userInput}
-        setUserInput={setUserInput}
-        taskList={taskList}
-        setTaskList={setTaskList}
+        addNewTask={(taskContent: string) => addNewTask(taskContent)}
       />
       <TaskDashboard taskList={taskList} />
     </div>
